Clarify language around side-effect definition
The side-effect of an instruction may or may not be defined by the ISA.
For example, the original RSA implementation had a side-effect of not
being constant-time, cycle-time, and instruction-retired-time with
different inputs. This meant that the value architecturally defined
counters for number of cycles and instruction retired would provide the
necesary information to recover the secret data the RSA was operating on.